# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/settings/notifications": {
      "get": {
        "operationId": "SettingsController_getNotificationSettings_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingsResponse"
                }
              }
            }
          }
        },
        "summary": "Read notification settings",
        "tags": ["settings"]
      },
      "patch": {
        "operationId": "SettingsController_updateNotificationSettings_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingsResponse"
                }
              }
            }
          }
        },
        "summary": "Update notification settings (partial)",
        "tags": ["settings"]
      }
    },
    "/v1/automation/jobs/unprocessed": {
      "get": {
        "operationId": "AutomationController_unprocessedJobs_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnprocessedJobsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "Fetch raw jobs awaiting LLM processing",
        "tags": ["automation"]
      }
    },
    "/v1/automation/jobs/dead-letter": {
      "get": {
        "operationId": "AutomationController_deadLetterJobs_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeadLetterJobResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "List raw jobs dead-lettered after repeated processing failures",
        "tags": ["automation"]
      }
    },
    "/v1/automation/jobs/{rawJobId}/results": {
      "post": {
        "operationId": "AutomationController_persistResult_v1",
        "parameters": [
          {
            "name": "rawJobId",
            "required": true,
            "in": "path",
            "description": "scraper.jobs_raw.id.",
            "schema": {
              "example": "42"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobResultDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultAckResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "Persist one processing-chain result",
        "tags": ["automation"]
      }
    },
    "/v1/automation/matches/unnotified": {
      "get": {
        "operationId": "AutomationController_unnotifiedMatches_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnnotifiedMatchResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "List matches awaiting notification",
        "tags": ["automation"]
      }
    },
    "/v1/automation/notifications": {
      "post": {
        "operationId": "AutomationController_recordNotification_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationRecordedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "Record a sent notification",
        "tags": ["automation"]
      }
    },
    "/v1/automation/digest": {
      "get": {
        "operationId": "AutomationController_digest_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "Jobs and matches new since the last digest",
        "tags": ["automation"]
      }
    },
    "/v1/automation/digest/sent": {
      "post": {
        "operationId": "AutomationController_markDigestSent_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestSentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "Advance the digest watermark to now",
        "tags": ["automation"]
      }
    },
    "/v1/automation/settings": {
      "get": {
        "operationId": "AutomationController_settings_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationSettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "internal-token": []
          }
        ],
        "summary": "Read the effective notification configuration for workflows",
        "tags": ["automation"]
      }
    }
  },
  "info": {
    "title": "job-hunter API",
    "description": "Gateway for jobs, reactions, keyword dictionaries, profiles, LLM admin, and sources.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TelegramSettingsResponse": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "chatId": {
            "type": "string",
            "nullable": true
          },
          "botTokenEnv": {
            "type": "string"
          },
          "botTokenConfigured": {
            "type": "boolean"
          }
        },
        "required": ["enabled", "chatId", "botTokenEnv", "botTokenConfigured"]
      },
      "EmailSettingsResponse": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "smtpHost": {
            "type": "string",
            "nullable": true
          },
          "smtpPort": {
            "type": "number",
            "nullable": true
          },
          "smtpUser": {
            "type": "string",
            "nullable": true
          },
          "smtpPasswordEnv": {
            "type": "string"
          },
          "smtpPasswordConfigured": {
            "type": "boolean"
          },
          "fromEmail": {
            "type": "string",
            "nullable": true
          },
          "toEmail": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "enabled",
          "smtpHost",
          "smtpPort",
          "smtpUser",
          "smtpPasswordEnv",
          "smtpPasswordConfigured",
          "fromEmail",
          "toEmail"
        ]
      },
      "NotificationSettingsResponse": {
        "type": "object",
        "properties": {
          "telegram": {
            "$ref": "#/components/schemas/TelegramSettingsResponse"
          },
          "email": {
            "$ref": "#/components/schemas/EmailSettingsResponse"
          },
          "matchThreshold": {
            "type": "number"
          },
          "digestHour": {
            "type": "number"
          }
        },
        "required": ["telegram", "email", "matchThreshold", "digestHour"]
      },
      "UpdateNotificationSettingsDto": {
        "type": "object",
        "properties": {
          "telegramEnabled": {
            "type": "boolean",
            "description": "Whether the Telegram channel is enabled."
          },
          "telegramChatId": {
            "type": "string",
            "description": "Destination Telegram chat id."
          },
          "telegramBotTokenEnv": {
            "type": "string",
            "description": "Name of the environment variable holding the Telegram bot token (never the value).",
            "example": "TELEGRAM_BOT_TOKEN"
          },
          "emailEnabled": {
            "type": "boolean",
            "description": "Whether the email channel is enabled."
          },
          "smtpHost": {
            "type": "string",
            "description": "SMTP server host."
          },
          "smtpPort": {
            "type": "number",
            "description": "SMTP server port (1-65535)."
          },
          "smtpUser": {
            "type": "string",
            "description": "SMTP username."
          },
          "smtpPasswordEnv": {
            "type": "string",
            "description": "Name of the environment variable holding the SMTP password (never the value).",
            "example": "SMTP_PASSWORD"
          },
          "fromEmail": {
            "type": "string",
            "description": "Sender address for the digest email."
          },
          "toEmail": {
            "type": "string",
            "description": "Recipient address for the digest email."
          },
          "matchThreshold": {
            "type": "number",
            "description": "Minimum match score (0-100) that triggers a notification."
          },
          "digestHour": {
            "type": "number",
            "description": "Hour of day (0-23) the daily digest is intended to run."
          }
        }
      },
      "LlmProfileInputResponse": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferences": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["summary", "skills", "preferences"]
      },
      "UnprocessedJobResponse": {
        "type": "object",
        "properties": {
          "rawJobId": {
            "type": "number",
            "example": 42
          },
          "sourceId": {
            "type": "number",
            "example": 1
          },
          "externalId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        },
        "required": ["rawJobId", "sourceId", "externalId", "url", "title", "body"]
      },
      "UnprocessedJobsResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/LlmProfileInputResponse"
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnprocessedJobResponse"
            }
          }
        },
        "required": ["profile", "jobs"]
      },
      "DeadLetterJobResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42
          },
          "sourceId": {
            "type": "number",
            "example": 1
          },
          "sourceSlug": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "processAttempts": {
            "type": "number",
            "example": 3
          },
          "processedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sourceId",
          "sourceSlug",
          "externalId",
          "url",
          "title",
          "processAttempts",
          "processedAt"
        ]
      },
      "JobResultStatus": {
        "type": "string",
        "enum": ["processed", "failed"]
      },
      "NormalizedJobDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "remote": {
            "type": "boolean",
            "nullable": true
          },
          "seniority": {
            "type": "string",
            "nullable": true
          },
          "salaryMin": {
            "type": "number",
            "nullable": true
          },
          "salaryMax": {
            "type": "number",
            "nullable": true
          },
          "salaryCurrency": {
            "type": "string",
            "nullable": true
          },
          "descriptionMd": {
            "type": "string"
          }
        },
        "required": ["title", "descriptionMd"]
      },
      "MatchDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "explanation": {
            "type": "string"
          },
          "modelUsed": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["score", "explanation"]
      },
      "CoverLetterDto": {
        "type": "object",
        "properties": {
          "bodyMd": {
            "type": "string"
          },
          "modelUsed": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["bodyMd"]
      },
      "JobResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/JobResultStatus"
              }
            ]
          },
          "sourceId": {
            "type": "number"
          },
          "externalId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "normalized": {
            "$ref": "#/components/schemas/NormalizedJobDto"
          },
          "match": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MatchDto"
              }
            ]
          },
          "coverLetter": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CoverLetterDto"
              }
            ]
          }
        },
        "required": ["status"]
      },
      "JobResultAckResponse": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/JobResultStatus"
              }
            ]
          },
          "jobId": {
            "type": "string",
            "nullable": true,
            "example": "123"
          }
        },
        "required": ["status", "jobId"]
      },
      "UnnotifiedMatchResponse": {
        "type": "object",
        "properties": {
          "jobMatchId": {
            "type": "string",
            "example": "17"
          },
          "jobId": {
            "type": "string",
            "example": "42"
          },
          "jobTitle": {
            "type": "string"
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "explanation": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["jobMatchId", "jobId", "jobTitle", "company", "url", "score", "explanation"]
      },
      "NotificationChannel": {
        "type": "string",
        "enum": ["telegram", "email"]
      },
      "RecordNotificationDto": {
        "type": "object",
        "properties": {
          "jobMatchId": {
            "type": "string",
            "description": "Job match id (bigint as string).",
            "example": "17"
          },
          "channel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationChannel"
              }
            ]
          }
        },
        "required": ["jobMatchId", "channel"]
      },
      "NotificationRecordedResponse": {
        "type": "object",
        "properties": {
          "jobMatchId": {
            "type": "string",
            "example": "17"
          },
          "channel": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationChannel"
              }
            ]
          }
        },
        "required": ["jobMatchId", "channel"]
      },
      "DigestJobSummaryResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "example": "42"
          },
          "title": {
            "type": "string"
          },
          "sourceSlug": {
            "type": "string"
          },
          "firstSeenAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["jobId", "title", "sourceSlug", "firstSeenAt"]
      },
      "DigestMatchSummaryResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "example": "42"
          },
          "title": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "url": {
            "type": "string"
          }
        },
        "required": ["jobId", "title", "score", "url"]
      },
      "DigestResponse": {
        "type": "object",
        "properties": {
          "since": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "newJobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DigestJobSummaryResponse"
            }
          },
          "newMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DigestMatchSummaryResponse"
            }
          }
        },
        "required": ["since", "newJobs", "newMatches"]
      },
      "DigestSentResponse": {
        "type": "object",
        "properties": {
          "lastDigestAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["lastDigestAt"]
      },
      "AutomationSettingsResponse": {
        "type": "object",
        "properties": {
          "telegramEnabled": {
            "type": "boolean"
          },
          "telegramChatId": {
            "type": "string",
            "nullable": true
          },
          "emailEnabled": {
            "type": "boolean"
          },
          "toEmail": {
            "type": "string",
            "nullable": true
          },
          "matchThreshold": {
            "type": "number"
          },
          "digestHour": {
            "type": "number"
          }
        },
        "required": [
          "telegramEnabled",
          "telegramChatId",
          "emailEnabled",
          "toEmail",
          "matchThreshold",
          "digestHour"
        ]
      }
    }
  }
}